+++ CHANGELOG.md
@@ -6,9 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
-- Added `Visualize` function to parse the graph in a container into DOT format
-  to allow visualization of errors types dependency relationship of types in
-  container.
+- Added `Visualize` function to visualize the state of the container in the
+  GraphViz DOT format. This allows visualization of error types and the
+  dependency relationships of types in the container.
 - Added `CanVisualizeError` function to determine if an error can be visualized
   in the graph.
 - Added `Name` option for `Provide` to add named values to the container