+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.4.0] - 2018-08-16
 ### Added
 - Added `Visualize` function to visualize the state of the container in the
   GraphViz DOT format. This allows visualization of error types and the
@@ -120,7 +120,7 @@ First release candidate.
 
 Initial release.
 
-[Unreleased]: https://github.com/uber-go/dig/compare/v1.3.0...HEAD
+[1.4.0]: https://github.com/uber-go/dig/compare/v1.3.0...v1.4.0
 [1.3.0]: https://github.com/uber-go/dig/compare/v1.2.0...v1.3.0
 [1.2.0]: https://github.com/uber-go/dig/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/uber-go/dig/compare/v1.0.0...v1.1.0